fix likes

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -13,6 +13,6 @@ public class Film {
     private String description;
     private LocalDate releaseDate;
     private int duration;
-    private Long likesCount;
+    private int likesCount;
     private List<Long> userIdLikes;
 }
